useLazyRef -> useRefWithInit
